TriggerCtrlMgr: matched getCurrentResActionIdx

+++ include/xlink2/xlink2TriggerCtrlMgr.h
@@ -36,8 +36,9 @@ public:
     void changeAction(s32, s32, s32);
     void setActionFrame(s32, s32);
     void stopAction(s32);
+    
     bool isCurrentActionNeedToObserve(s32) const;
-    u64 getCurrentResActionIdx(s32) const;
+    s32 getCurrentResActionIdx(s32) const;
 
     void updateGlobalPropertyTriggerCtrl();
